build: apply warnings-as-errors per target and clear MSVC /W4 /WX
The CI warning gate injected warning flags via CMAKE_CXX_FLAGS, which on MSVC
replaced the compiler defaults (dropping /EHsc and breaking exception unwinding in
STL headers) and held third-party headers (units) to this project's warning level.
Rework it so it is correct and portable:

- Add a COORDINATES_WARNINGS_AS_ERRORS option and a coordinates_set_warnings()
  helper that applies the warning level and -Werror//WX per target via
  target_compile_options, preserving compiler defaults. CI now sets the option
  instead of clobbering CMAKE_CXX_FLAGS.
- Mark the units headers as system includes (INTERFACE_SYSTEM_INCLUDE_DIRECTORIES +
  /external) so units' own warnings never fail our build.
- Calibrate MSVC /W4: /wd4459 (a local hiding a library global single-letter name,
  stylistic) and /wd4702 (unreachable code emitted by the optimizer inside units'
  constexpr pow template, a third-party header /external:W0 cannot suppress);
  _CRT_SECURE_NO_WARNINGS for the deliberate std::getenv (C4996). Our own
  unreachable/dead code stays covered by -Wall -Werror on the GCC/Clang legs.
- Fix genuine MSVC narrowings in our code: static_cast the double->unsigned and
  size_t->int conversions in dtedTile.cpp / geoid.h; use EXPECT_EQ (integer) instead
  of EXPECT_DOUBLE_EQ for tile-count comparisons in topographyTest.h; make the
  hillshade kernel locals z-prefixed to avoid shadowing.
- image()/hillshade() in algorithm.h are inline (header-defined public helpers).

Both warning gates are now clean: MSVC cl.exe /W4 /WX and g++15 -Werror -Wall
-Wextra each build with zero warnings; full suite 460/460.

+++ src/dtedTile.cpp
@@ -216,8 +216,8 @@ meters<> DTEDTile::elevation(const degrees<>& latitude, const degrees<>& longitu
 	double lonY = (longitude - m_metadata.southwestLongitude()) / m_metadata.longitudeResolution();
 
 	// find the lat/lon index (integer)
-	unsigned int latIndex = floor(latX);
-	unsigned int lonIndex = floor(lonY);
+	unsigned int latIndex = static_cast<unsigned int>(floor(latX));
+	unsigned int lonIndex = static_cast<unsigned int>(floor(lonY));
 
 	// interpolation x,y
 	double xVal = (latX - latIndex);
